Modularize jira/mod.rs into submodules for maintainability
Split the 778-line jira/mod.rs into smaller focused modules:
- link.rs: JIRA link flow (~200 lines)
- forge_impl.rs: Forge trait implementation (~400 lines)
- mod.rs: Configuration and helpers (~180 lines)
